move statistics code to separate helper file
